test/index.js: fix mocha test problem

+++ test/index.js
@@ -5,7 +5,8 @@ const { exec } = require("child_process");
 describe('svg2pdf', function() {
     it('should correctly convert svg file to pdf file', function() {
         cmd = 'node ./bin/index.js -w 50 -h 50 ./test/test.svg ./test/test.pdf';
-        fs.unlinkSync('./test/test.pdf');
+        if (fs.existsSync('./test/test.pdf'))
+            fs.unlinkSync('./test/test.pdf');
         assert.equal(fs.existsSync('./test/test.pdf'), false);
         exec(cmd, (error, stdout, stderr) => {
             if (error) {
